# spec/contracts/modeling/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Modeling API
  version: 1.0.0
  description: |
    Синхронный REST API для BC Modeling (docs/backend.md §5).

    Каждая операция ссылается на use case в `x-use-case` и на операцию таблицы
    доступа в `x-operation` (spec/nfr/modeling.yaml#security.authorization): роль
    проверяется на границе use case'а, а не матчером URL. Наполнено на этапе 2
    (2026-10-08) по одному endpoint'у на экспонированный сценарий; до этого
    `paths` был пуст намеренно — endpoint без `UC-*` нарушал бы трассируемость.

    Ошибки — `application/problem+json` (§8.1) с полем `code`: код инварианта
    (`INV-MDL-NNN`) или бизнес-ошибки (`RELATION_NOT_PERMITTED`, `MDL_*`).
    Класс отказа домена определяет статус: не найдено или скрыто списком доступа —
    `404`, конфликт состояния — `409`, нарушение правила — `422`, роль — `403`,
    очищенный снимок — `410`.

    Остаток XML (`raw_xml`, ADR-0017) наружу не отдаётся: это внутренний формат
    interchange.
servers:
  - url: /api/v1
    description: Относительный путь; приложение и фронтенд отдаются одним артефактом.
security:
  - bearerAuth: []
...
paths:
...
  /models:
    get:
      tags: [models]
      operationId: listModels
      summary: Модели пространства, видимые автору (FR-04, INV-MDL-011)
      x-use-case: UC-MDL-001
      x-operation: OpenModel
      parameters:
        - { in: query, name: workspaceId, required: false, schema: { type: string, format: uuid },
            description: "Без него — первое пространство." }
      responses:
        '200':
          description: Удалённые видит только ADMIN.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ModelSummary' } }
    post:
      tags: [models]
      operationId: createModel
      summary: Пустая модель с девятью корневыми папками и версией 1
      x-use-case: UC-MDL-001
      x-operation: CreateModel
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, maxLength: 500 }
                workspaceId: { type: string, format: uuid }
      responses:
        '201':
          description: Модель создана; повтор ключа с тем же телом — та же модель (INV-MDL-003).
          content:
            application/json: { schema: { $ref: '#/components/schemas/ModelSummary' } }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '409': { $ref: '#/components/responses/Conflict409' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT, выданный Keycloak. Роли VIEWER / ARCHITECT / ADMIN — в claim'ах токена (FR-28), группы — в `groups`.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, maxLength: 64 }
      description: |
        Ключ дедупликации повторов. Тот же ключ с тем же телом возвращает результат
        первого выполнения; тот же ключ с другим телом — `409` (INV-MDL-003).
        Ключ принадлежит пользователю: чужой ключ с тем же значением не конфликтует.
...
  schemas:
    Problem:
      type: object
      description: |
        RFC 7807 Problem Details + поле `code` (docs/backend.md §8.1). Дополнительные поля
        по ситуации: `lockOwner` и `expiresAt` у чужой блокировки, `relationships` у удаления
        связанного элемента, `permitted` у недопустимой связи.
      required: [type, title, status, code]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        code:
          type: string
          description: Код инварианта (`INV-MDL-NNN`, domain/modeling/invariants.md) или иной бизнес-ошибки.
          examples: ["INV-MDL-004", "INV-MDL-006", "RELATION_NOT_PERMITTED"]
        lockOwner:     { type: string }
        relationships: { type: array, items: { type: string, format: uuid } }
        permitted:     { type: array, items: { type: string } }
...
    ModelSummary:
      type: object
      properties:
        id:            { type: string, format: uuid }
        workspaceId:   { type: string, format: uuid }
        archiId:       { type: string }
        name:          { type: string }
        documentation: { type: string }
        status:        { type: string, enum: [ACTIVE, DELETED] }
        archiVersion:  { type: string }
        createdBy:     { type: string }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
...
  responses:
...
    Forbidden403:
      description: Роль не даёт права на операцию (FR-28) либо список доступа даёт только чтение.
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    Conflict409:
      description: |
        Блокировки нет или она чужая, конкурентное сохранение (INV-MDL-006), повтор
        Idempotency-Key с другим телом (INV-MDL-003), удаление связанного элемента
        (INV-MDL-004), команда к удалённой модели (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    UnprocessableEntity422:
      description: |
        Нарушение бизнес-правила: недопустимая связь (INV-MDL-007), тип не той фазы,
        папка чужой модели или не того корня (INV-MDL-009), запрещённый переход (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
...